UIKIT-188: feat(ui,MaskField): исправления
